backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorgram.py==1.2.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
macholib==1.16.3
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.97.0
openpyxl==3.1.5
orjson==3.11.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pdf2image==1.17.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pygame==2.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyMuPDF==1.25.4
PyPDF2==3.0.1
pyreadstat==1.2.8
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.27.0
stripe==8.5.0
tqdm==4.67.1
ttkthemes==3.2.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
